ST6RI-751 Consolidated subject and object utility methods in UsageUtil.
