+++ org.omg.sysml/src/org/omg/sysml/adapter/RequirementUsageAdapter.java
@@ -1,6 +1,6 @@
 /*******************************************************************************
  * SysML 2 Pilot Implementation
- * Copyright (c) 2021, 2023 Model Driven Solutions, Inc.
+ * Copyright (c) 2021, 2023-2024 Model Driven Solutions, Inc.
  *    
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -30,7 +30,6 @@ import org.omg.sysml.lang.sysml.RequirementDefinition;
 import org.omg.sysml.lang.sysml.RequirementUsage;
 import org.omg.sysml.lang.sysml.Type;
 import org.omg.sysml.lang.sysml.Usage;
-import org.omg.sysml.util.TypeUtil;
 import org.omg.sysml.util.UsageUtil;
 
 public class RequirementUsageAdapter extends ConstraintUsageAdapter {
@@ -80,17 +79,17 @@ public class RequirementUsageAdapter extends ConstraintUsageAdapter {
 	@Override
 	protected List<? extends Feature> getRelevantFeatures(Type type, Element skip) {
 		return UsageUtil.isObjective(getTarget())? 
-				Collections.singletonList(TypeUtil.getObjectiveRequirementOf(type)):
+				Collections.singletonList(UsageUtil.getObjectiveRequirementOf(type)):
 			    super.getRelevantFeatures(type, skip);
 	}
 	
 	// Transformation
-
+	
 	@Override
 	public void doTransform() {
 		RequirementUsage target = getTarget();
 		super.doTransform();
-		computeSubjectParameterOf(target);
+		UsageUtil.addSubjectParameterTo(target);
 	}
 	
 }
