+++ org.omg.sysml/src/org/omg/sysml/delegate/RequirementUsage_subjectParameter_SettingDelegate.java
@@ -25,7 +25,7 @@ import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.EStructuralFeature;
 import org.eclipse.emf.ecore.InternalEObject;
 import org.omg.sysml.lang.sysml.RequirementUsage;
-import org.omg.sysml.util.TypeUtil;
+import org.omg.sysml.util.UsageUtil;
 
 public class RequirementUsage_subjectParameter_SettingDelegate extends BasicDerivedObjectSettingDelegate {
 
@@ -35,7 +35,7 @@ public class RequirementUsage_subjectParameter_SettingDelegate extends BasicDeri
 
 	@Override
 	protected EObject basicGet(InternalEObject owner) {
-		return TypeUtil.basicGetSubjectParameterOf((RequirementUsage)owner);
+		return UsageUtil.basicGetSubjectParameterOf((RequirementUsage)owner);
 	}
 
 }
