+++ org.omg.sysml/src/org/omg/sysml/util/TypeUtil.java
@@ -34,13 +34,9 @@ import java.util.stream.Stream;
 import org.eclipse.emf.common.util.BasicEList;
 import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.ecore.EClass;
-import org.omg.sysml.adapter.DefinitionAdapter;
 import org.omg.sysml.adapter.TypeAdapter;
 import org.omg.sysml.lang.sysml.Association;
 import org.omg.sysml.lang.sysml.BindingConnector;
-import org.omg.sysml.lang.sysml.CaseDefinition;
-import org.omg.sysml.lang.sysml.CaseUsage;
-import org.omg.sysml.lang.sysml.Definition;
 import org.omg.sysml.lang.sysml.Element;
 import org.omg.sysml.lang.sysml.Expression;
 import org.omg.sysml.lang.sysml.Feature;
@@ -52,13 +48,10 @@ import org.omg.sysml.lang.sysml.Membership;
 import org.omg.sysml.lang.sysml.OccurrenceDefinition;
 import org.omg.sysml.lang.sysml.OccurrenceUsage;
 import org.omg.sysml.lang.sysml.ParameterMembership;
-import org.omg.sysml.lang.sysml.RequirementUsage;
 import org.omg.sysml.lang.sysml.ReturnParameterMembership;
-import org.omg.sysml.lang.sysml.SubjectMembership;
 import org.omg.sysml.lang.sysml.SysMLFactory;
 import org.omg.sysml.lang.sysml.SysMLPackage;
 import org.omg.sysml.lang.sysml.Type;
-import org.omg.sysml.lang.sysml.Usage;
 
 public class TypeUtil {
 	
@@ -445,28 +438,6 @@ public class TypeUtil {
 				collect(Collectors.toList());
 	}
 
-	// Subject parameters
-
-	public static Feature getSubjectParameterOf(Type type) {
-		return type instanceof Definition? ((DefinitionAdapter)ElementUtil.getElementAdapter((Definition)type)).getSubjectParameter():
-			   type instanceof Usage? UsageUtil.getUsageAdapter((Usage)type).getSubjectParameter():
-			   null;
-	}
-
-	public static Usage basicGetSubjectParameterOf(Type type) {
-		ElementUtil.transform(type);
-		return (Usage)getOwnedFeatureByMembershipIn(type, SubjectMembership.class);
-	}
-	
-	// Objective requirements
-
-	public static RequirementUsage getObjectiveRequirementOf(Type type) {
-		ElementUtil.transform(type);
-		return type instanceof CaseDefinition? ((CaseDefinition)type).getObjectiveRequirement():
-			   type instanceof CaseUsage? ((CaseUsage)type).getObjectiveRequirement():
-			   null;
-	}
-	
 	// Associations
 
 	public static Type getSourceTypeOf(Association association) {
